Starting Files
This is beginning state of the MyTunes tutorial.   I have included an
index.html file with a starting stylesheet and blank JQuery program
file.   I have included 3 PHP utiliities for using JSON and retreiving
image BLOBs from the database.   Generally, it is not considered good
practice to store images in BLOBs; I do so in this tutorial to show how
it is done, you can decide in what cases to use this technique.   In
terms of maintenance and network, the data is basically thumbnail sized
and not overly cumbersome.   I have included the base JPG files also.

The database contains the 4 tables we have already explored in our
class - genres - artists - albums - tracks.   You should be able to
import the database, including the image data, by running the scripts
in /db in that order.

The Bootstrap files are also included here and referenced on our page,
we are ready to start programming - wiring our database information to
the web document and building DOM elements to display the data as
information.