First Basic Database Interaction
Our first AJAX programming is basic, we create a Javascript function in
our mytunes.js called getMusic( ), which makes an AJAX GET request.
It is expecting data to be returned in the JSON format.   In addition,
we turn OFF the default asynchronous processing of AJAX  - which is
required in this case because we are returning data to a variable in
the outer scope of the JQuery wrapper from within the .done( ) event
handler function.

We call our function by assigning its return value to the variable with
broad scope.   Then we log the data to the browser console.

The get-music.php server requires a MySQL data connection in a file the
I DID NOT PROVIDE on this repository for security reasons.   The file
should simple return a connection called mysqli using the mysqli PHP
object.

The connection then opens a new query selecting every column from every
row in the genres table and creating an object indexed by the idgenres
column.

The student server does not support JSON natively - the version of PHP
is out of date.   I have provided a JSON service file in /php/utils and
we then call a new JSON object and echo that back to the AJAX request.

The results are not visibile on the page, but you should see an object
with 5 genre objects listed in the Console.