Completing the JSON structure
Now we add the next two table queries to complete the JSON structure
that represents the database.   While we haven't added any of this to a
visible page DOM element yet, we have the complete database information
in memory now in our browser to reference.

Actually, not completely, complete.    We did not request the image
data from the artists and albums tables, we use another technique to so
do when we build our display elements.