Display genres in a Tab navigation
We'll list our music genres in a tab navigation.    To do so, we create
a new Javascript function - the showGenres( ) function builds a string
of HTML data based on the Bootstrap Tab navigation structure.  We
insert two data elements from the Javascript object holding our music
library data - the idgenres primary key is added to the href attribute
and the genre name is placed in the tab text.

The JQuery .each( ) function iterates through all the genres in the
database returned by the get-music query and builds the navigation list.

We then add the content to the #genre element on our HTML page and make
the first tab active.

Next we call the showGenres( ) function to invoke it from the main
program flow of the Javascript program.