Display artists for each genre in tab panes
A new function builds HTML content in much the same way, by
concatenating a string inserting information from our data object into
specific elements such as the idgenres primary key in the id of the
pane - to coordinate the tab navigation correctly.   We then list each
artist in an element with class hero-unit and span4, wrapped by the tab
pane.

For each artist we list their name and we also display the image.
Note how the src attribute of the image element calls a PHP uri with a
query string of the idartists primary key from the Javascript element
(through the .each( ) function.)

This content is then appended into the tab structure as the collection
of panes, each with the set of artists for the pane's genre, and the
first one is activated.

On line 23 - the new function is called - from within the showGenres( )
function - anytime we show Genres, we'll want to add the artists to
display in the tab panes.