List the albums for each artist
A new function builds the HTML content for a display block that lists
the artist's albums and is aligned to the right of the of the artist
hero-unit.    Again, the same technique is used to retrieve the image.
  The PHP program returns just the image data from the database in
image content type.

We place this block into the DOM by calling the function from within
showArtist( ), passing both the indices in the object chain needed to
navigate to the albums for each artist.

We also add the simple CSS styling to make a grid of album cover and
title choices.