Adding new genres
We add a modal display pane to our index file and then in our
Javascript porgram we update the content generators to add a 'static'
tab to the menu with a plus icon and a data-toggle attribute used to
show the modal.

We also add a click event listener on the Save button of the modal to
call a new PHP program that will insert the new genre.

The .done( ) function of the AJAX call also updates the displays by
re-fetching the music object from the database and calling the
showGenres( ) function.