Added a simplified database creation script with no BLOBs
This one script creates all four tables for the MyTunes tutorial.   The
script does not include BLOBS, so you need to load the image
information using MySQL Workbench.   There is a context menu on the
column cell that allows you to upload data from a file, such as  a JPG.