Add New Track
We add a modal panel to the page to enable the user to add a new track.

In the Javascript code, we add a header to the Tracks section with
link element to display the modal.     Then most of the new behavior is
implemented in two click event handlers, one on the PLUS icon to show
the modal, and one on the Submit button to process the request.

See how we provide the parent structure data to the various functions
using data attributes to tie the data to the DOM elements.   Having the
parent data simplifies our ability to manage the data in the MUSIC
Javascript data object.

We add a new PHP service called new-track.php, which  is able to insert
the new track information with the album ID and the new Track name.
The MySQL autoincrement function assigns a new track ID as the primary
key and the PHP mysqli interface enables us to grab that ID in SQL
statement object - how cool is that!

Once the database insert is complete, we assemble a new Track oject
using Javascript literal notation and add that to the appropriate album
in the data object.

We are then able to call the showTracks( ) function to refresh the
tracks section of the page.     At this point, the data is already in
the memory of the browser, we do now need to request the data from the
database.