Rebuild the track listing as an interactive grid
First, we add an SQL statement to alter the TRACKS table to add a
column to store the address of the audio file.

We also update the showTracks( ) function to change the DOM development
of the track article element.    We add update and delete buttons for
each track.    We use a contenteditable attribute on the data elements
to allow for updates.

Note how the article holds the trackid data for each row.  After the
DOM is built in showTracks, we add click event listeners for the
buttons.   Now how we are able to retreive the trackids relative to
each track using the parent( ) jQuery function.